requirements.txt
langgraph>=0.0.19
langchain>=0.0.335
langchain-groq>=0.0.2
groq>=0.4.1
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
imbalanced-learn>=0.10.0
python-dotenv>=1.0.0
